requirements.txt
Pillow==9.2.0
opencv-python==4.5.5.64
numpy==1.22.4
mysql-connector-python==8.0.29
opencv-contrib-python==4.6.0.66
